Move `stylelint-vscode` from `/lib` to `/server`
