+++ .vscodeignore
@@ -8,3 +8,4 @@
 !server.js
 !LICENSE.txt
 !lib/**/*
+!server/**/*
